requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
dotenv==0.0.5
frozendict==2.3.10
gunicorn==23.0.0
idna==3.10
Markdown==3.7
packaging==24.2
psycopg2==2.9.10
python-dateutil==2.7.5
python-dotenv==1.0.1
requests==2.32.3
resend==2.4.0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.3.0
tzdata==2024.2
urllib3==1.26.20
